+++ CHANGELOG.md
@@ -1,4 +1,6 @@
+* Added JsonApi action type
 * Improved OPTIONS request handling
+* Added matching route to Request attributes
 
 ## v0.1.3 (2023-12-11)
 * Improved JSON error handling